Abstract screen output

+++ src/fragment.rs
@@ -3,7 +3,7 @@ use crate::Color;
 pub struct Fragment {
     pub color: Color,
 
-    pub x: i32, // pixel coordinate
-    pub y: i32, // pixel coordinate
-    pub z: f64, // z-depth in NDC, normalized by [-1, 1]
+    pub x: usize, // pixel coordinate
+    pub y: usize, // pixel coordinate
+    pub z: f64,   // z-depth in NDC, normalized by [-1, 1]
 }
